Control settings for the zcurve EM algorithm

All these settings are passed to the Expectation Maximization fitting algorithm. All unspecified settings are set to the default value. Setting model = "EM" sets all settings to the default value irrespective of any other setting and fits z-curve as described in Bartoš and Schimmack (2020)

Arguments

model

A type of model to be fitted, defaults to "EM" for a z-curve with 7 z-scores centered components.

sig_level

An alpha level of the test statistics, defaults to .05

a

A beginning of fitting interval, defaults to qnorm(sig_level/2,lower.tail = F)

b

An end of fitting interval, defaults to 5

mu

Means of the components, defaults to 0:6

sigma

A standard deviation of the components, defaults to rep(1, length(mu))

theta_alpha

A vector of alpha parameters of a Dirichlet distribution for generating random starting values for the weights, defaults to rep(.5, length(mu))

theta_max

Upper limits for weights, defaults to rep(1,length(mu))

criterion

A criterion to terminate the EM algorithm, defaults to 1e-6

criterion_start

A criterion to terminate the starting phase of the EM algorithm, defaults to 1e-3

criterion_boot

A criterion to terminate the bootstrapping phase of the EM algorithm, defaults to 1e-5

max_iter

A maximum number of iterations of the EM algorithm (not including the starting iterations) defaults to 10000

max_iter_start

A maximum number of iterations for the starting phase of EM algorithm, defaults to 100

max_iter_boot

A maximum number of iterations for the booting phase of EM algorithm, defaults to 100

fit_reps

A number of starting fits to get the initial position for the EM algorithm, defaults to 100

References

Bartoš F, Schimmack U (2020). “Z-curve. 2.0: Estimating Replication Rates and Discovery Rates.” doi:10.31219/osf.io/wr93f , submitted for publication.

See also

Examples

# to increase the number of starting fits
# and change the means of the mixture components

ctrl <- list(
   fit_reps  = 50,
   mu = c(0, 1.5, 3, 4.5, 6)
)
if (FALSE) zcurve(OSC.z, method = "EM", control = ctrl)