Control settings for the original z-curve density algorithm

All settings are passed to the density fitting algorithm. All unspecified settings are set to the default value. Setting model = "KD1" sets all settings to the default value irrespective of any other setting and fits z-curve as described in Brunner and Schimmack (2020) .

Arguments

version
Set to 1 to fit the original version of z-curve. Defaults to 2 = the updated version of z-curve. For its settings page go to control_density.

model
A type of model to be fitted, defaults to "KD1" (the only possibility)

sig_level
An alpha level of the test statistics, defaults to .05

a
A beginning of fitting interval, defaults to qnorm(sig_level/2,lower.tail = F)

b
An end of fitting interval, defaults to 6

K
Number of mixture components, defaults to 3

max_iter
A maximum number of iterations for the nlminb optimization for fitting mixture model, defaults to 150

max_eval
A maximum number of evaluation for the nlminb optimization for fitting mixture model, defaults to 300

criterion
A criterion to terminate nlminb optimization, defaults to 1e-10

bw
A bandwidth of the kernel density estimation, defaults to "nrd0"

References

Brunner J, Schimmack U (2020). “Estimating population mean power under conditions of heterogeneity and selection for significance.” Meta-Psychology, 4. doi: 10.15626/MP.2018.874 .

See also

Examples

# to increase the number of iterations
ctrl <- list(
   version   = 1,
   max_iter  = 300
)
zcurve(OSC.z, method = "density", control = ctrl)

#> Call:
#> zcurve(z = OSC.z, method = "density", control = ctrl)
#> 
#> Estimates:
#>       ERR       EDR 
#> 0.6370137 0.3718363