Plot fitted z-curve object

Plot fitted z-curve object

# S3 method for zcurve
plot(
  x,
  annotation = FALSE,
  CI = FALSE,
  extrapolate = FALSE,
  plot_type = "base",
  y.anno = c(0.95, 0.88, 0.78, 0.71, 0.61, 0.53, 0.43, 0.35),
  x.anno = 0.6,
  cex.anno = 1,
  ...
)

Arguments

x

Fitted z-curve object

annotation

Add annotation to the plot. Defaults to FALSE.

CI

Plot confidence intervals for the estimated z-curve. Defaults to FALSE.

extrapolate

Scale the chart to the extrapolated area. Defaults to FALSE.

plot_type

Type of plot to by produced. Defaults to "base" for th base plotting function. An alternative is "ggplot" for a ggplot2.

y.anno

A vector of length 8 specifying the y-positions of the individual annotation lines relative to the figure's height. Defaults to c(.95, .88, .78, .71, .61, .53, .43, .35)

x.anno

A number specifying the x-position of the block of annotations relative to the figure's width.

cex.anno

A number specifying the size of the annotation text.

...

Additional arguments including main, xlab, ylab, xlim, ylim, cex.axis, cex.lab

See also

Examples

if (FALSE) {
# simulate some z-statistics and fit a z-curve
z <- abs(rnorm(300,3))
m.EM <- zcurve(z, method = "EM", bootstrap = 100)

# plot the z-curve
plot(m.EM)

# add annotation text and model fit CI
plot(m.EM, annotation = TRUE, CI = TRUE)

# change the location of the annotation to the left
plot(m.EM, annotation = TRUE, CI = TRUE, x_text = 0)
}