Fit a z-curve to clustered data

zcurve_clustered is used to fit z-curve models to clustered data. The function requires a data object created with the zcurve_data() function as the input (where id denotes clusters). Two different methods that account for clustering ar implemented via the EM model: "w" for down weighting the likelihood of the test statistics proportionately to the number of repetitions in the clusters, and "b" for a nested bootstrap where only a single study from each bootstrap is selected for model fitting.

zcurve_clustered(
  data,
  method = "b",
  bootstrap = 1000,
  parallel = FALSE,
  control = NULL
)

Arguments

data

an object created with zcurve_data() function.

method

the method to be used for fitting. Possible options are down weighting "w" and nested bootstrap "b". Defaults to "w".

bootstrap

the number of bootstraps for estimating CI. To skip bootstrap specify FALSE.

parallel

whether the bootstrap should be performed in parallel. Defaults to FALSE. The implementation is not completely stable and might cause a connection error.

control

additional options for the fitting algorithm more details in control EM.

Value

The fitted z-curve object

References

There are no references for Rd macro \insertAllCites on this help page.

See also